requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
defusedxml==0.7.1
Django==4.1.3
django-allauth==0.58.2
django-bootstrap-datepicker-plus==5.0.3
django-bootstrap4==22.3
django-crispy-forms==1.14.0
django-filter==22.1
idna==3.4
oauthlib==3.2.2
pycparser==2.21
pydantic==1.10.4
PyJWT==2.6.0

python-dotenv==0.21.1
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
soupsieve==2.3.2.post1
sqlparse==0.4.3
typing_extensions==4.4.0
urllib3==1.26.13
virtualenv==20.24.6
virtualenvwrapper==4.8.4
